+++ src/pd_ocr_synth/render/line.py
@@ -99,7 +99,7 @@ def render_line(
         raise MissingGlyphError(text, font.path, missing)
 
     handles = ctx.font_handles(font.path)
-    pixel_size = max(1, int(round(font_size_pt * dpi / 72.0)))
+    pixel_size = max(1, round(font_size_pt * dpi / 72.0))
     handles.ft_face.set_pixel_sizes(pixel_size, pixel_size)
 
     info_glyphs, positions = _shape(handles.hb_face, text, pixel_size, font.features)
@@ -163,7 +163,7 @@ def _composite_line(
         y_offset = pos.y_offset / 64.0
         x = pen_x + x_offset + bm["left"]
         y = pen_y - y_offset - bm["top"]
-        placements.append((bm, int(round(x)), int(round(y)), info.cluster))
+        placements.append((bm, round(x), round(y), info.cluster))
         pen_x += pos.x_advance / 64.0
         pen_y += pos.y_advance / 64.0
